Refactor to faster globs
